METER_CNN/getFirebase/up-get-firebase/requirements.txt
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
google-api-core==1.9.0
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.9
idna==2.8
nose==1.3.7
numpy==1.13.3
oauthlib==2.0.6
pandas==0.22.0
pbr==5.1.3
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
PySocks==1.6.8
python-dateutil==2.7.2
python-firebase==1.2
pytz==2019.1
requests==2.21.0
requests-oauthlib==0.8.0
rsa==4.0
six==1.12.0
stevedore==1.30.1
urllib3==1.24.1
virtualenv==16.4.3
virtualenv-clone==0.5.2
virtualenvwrapper==4.8.4
